Implementation of the about dialog in JavaFX
The old dialog and html layout file have been removed. The layout of the
dialog panel is defined in the AboutDialogLayout.fxml located in the
resources with its stylesheet. The AboutDialogController class takes
care of the text initialization of any components in the layout. It also
implements any methodes called from any component of the dialog panel.